Comunicacao feita e correcao logica de modo corrente e tratamento das informacoes recebidas

+++ Inc/com_protocol.h
@@ -36,11 +36,17 @@ typedef struct
 
 void COM_Protocol_Receive_Communication_Control(StatusMenssageTypeDef *Status_Message, 
 Load *Message_Received);
+
 void COM_Protocol_Transceiver_Communication_Control(StatusMenssageTypeDef *Status_Message, 
 uint8_t  *Message_To_Communication);
+
 StatusMenssageTypeDef COM_Protocol_Check_Menssage(uint8_t *data, uint32_t length);
+
 bool CHECKSUM(uint8_t *data, uint8_t length);
+
 Load Convert_Received_Serial_Message_To_Load_State(uint8_t Received_Message[]);
-uint16_t Convert_Data1_And_Data2_to_uint16_t(uint8_t Received_Datas[]);
+
+float Convert_Data1_And_Data2_to_uint16_t(uint8_t Received_Datas[]);
+
 uint8_t COM_Protocol_Report_Erro(StatusMenssageTypeDef Erro);
 #endif /* __COM_PROTOCOL_H */
