+++ Inc/main.h
@@ -51,7 +51,7 @@ typedef enum {
 
 typedef struct {
   E_Carga_State state_load;
-  uint16_t value_state_load;
+  float value_state_load;
   uint8_t time_load_on;
 } Load;
 /* USER CODE END EC */
